requirements.txt
antlr4-python3-runtime==4.13.2
ruamel.yaml >= 0.16.0, < 1.0.0
requests >= 2.24.0, < 3.0.0
cryptography >= 38.0.4
python-dotenv >= 1.0.0
google-generativeai >= 0.5.4
google-genai >= 1.32.0